+++ djangojs/conf.py
@@ -22,6 +22,7 @@ DEFAULTS = {
     'JS_URLS_UNNAMED': False,
     'JS_CONTEXT': None,
     'JS_CONTEXT_EXCLUDE': None,
+    'JS_CONTEXT_PROCESSOR': 'djangojs.utils.ContextSerializer',
     'JS_I18N_APPS': None,
     'JS_I18N_APPS_EXCLUDE': None,
     'JQUERY_VERSION': JQUERY_DEFAULT_VERSION,