+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Current
   - ``settings.JS_USER_ENABLED``
   - ``settings.JS_CONTEXT_ENABLED``
 - Added context black and white lists (``settings.JS_CONTEXT`` and ``settings.JS_CONTEXT_EXCLUDE``)
+- Allow context serialization customization by inheritance with ``settings.JS_CONTEXT_PROCESSOR``